@@ -1,10 +1,7 @@
 package com.hazukie.scheduleviews.fileutil;
 
-import android.util.Log;
-
 import com.google.gson.Gson;
 
-import java.lang.reflect.Type;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -13,24 +10,34 @@ public class Converter {
     public Converter(){
         if(gson==null) gson=new Gson();
     }
+
     public List<String> convertObj2Jsn(List<Object> objectList){
         List<String> stringList=new ArrayList<>();
-
         if(objectList.size()!=0){
             for(Object obj:objectList) stringList.add(gson.toJson(obj));
         }
         return stringList;
     }
 
-    public String convertObj2Jsn(Object obj){
+    public <T> String convertObj2Jsn(T  t,String appendStr){
         String s="";
-        if(obj!=null) s=gson.toJson(obj);
+        if(t!=null) s=gson.toJson(t)+appendStr;
         return s;
     }
 
-    public Object convertJsn2Obj(String content, Type type){
-        if(content.length()>0&&type!=null)
-        return gson.fromJson(content,type);
-        return "{}";
+    public <T> String convertObj2Jsn(T  t){
+        return convertObj2Jsn(t,"");
+    }
+
+    /**
+     *
+     * @param content 解释字符
+     * @param type 反射类型
+     * @return @null 返回值可能为空
+     * @param <T> 泛型
+     */
+    public <T> T convertJsn2Any(String content, Class<T> type){
+        if(content.length()>0&&type!=null) return gson.fromJson(content,type);
+        return null;
     }
 }
