@@ -8,6 +8,7 @@ import com.hazukie.scheduleviews.models.ScheWithTimeModel;
 import com.hazukie.scheduleviews.models.TimeHeadModel;
 import com.hazukie.scheduleviews.models.TimeModel;
 import com.hazukie.scheduleviews.models.Unimodel;
+import com.hazukie.scheduleviews.statics.Statics;
 import com.hazukie.scheduleviews.utils.CycleUtil;
 
 import java.util.ArrayList;
@@ -15,7 +16,7 @@ import java.util.List;
 
 public class OftenOpts extends Fileystem {
     private final Context context;
-    //private Fileystem fileystem;
+
     private static final String default_sch="空表";
     private static final String default_time="默认作息表";
 
@@ -27,33 +28,32 @@ public class OftenOpts extends Fileystem {
     private OftenOpts(Context context){
         super(context);
         this.context=context;
-        //if(fileystem==null) fileystem=Fileystem.getInstance(context);
     }
 
    public static OftenOpts getInstance(Context context) {
         if(instance==null) instance=new OftenOpts(context);
         return instance;
     }
+
     /*------总表方法------*/
     //获取课表数据
     public List<ClassLabel> getClsList(String name){
         List<ClassLabel> lis=new ArrayList<>();
-        List<Object> currentLis=getDataList(FileRootTypes.sches,name,ClassLabel.class);
-
-        if(currentLis.size()>0&&currentLis.get(0)!=null)
-            CycleUtil.cycle(currentLis,(obj, objects) -> lis.add((ClassLabel) obj));
+        List<ClassLabel> currentLis=getDataList(FileRootTypes.sches,name,ClassLabel.class);
+        if(currentLis.size()>0&&currentLis.get(0)!=null) return currentLis;
         return lis;
     }
 
-    //获取课表数据
-    public List<Object> getRawClsList(String name){
-        return getDataList(FileRootTypes.sches,name,ClassLabel.class);
-    }
-
     public boolean putRawClsList(String file_name,List<ClassLabel> classLabels){
-        if(classLabels.size()>0&&classLabels.get(0)!=null)
-            return putDatazList(FileRootTypes.sches,file_name,new ArrayList<>(classLabels));
-        else return false;
+        boolean check=true;
+        try{
+            if(classLabels.size()>0&&classLabels.get(0)!=null)
+                putDataList(FileRootTypes.sches,file_name,new ArrayList<>(classLabels));
+        }catch (Exception e){
+            check=false;
+            e.printStackTrace();
+        }
+        return check;
     }
 
     //写入索引数据
@@ -63,7 +63,14 @@ public class OftenOpts extends Fileystem {
 
     //写入数据
     public boolean putRawSctzList(List<ScheWithTimeModel> scts){
-        return putDatazList(FileRootTypes.index,default_sche_index_file_name,new ArrayList<>(scts));
+        boolean check=true;
+        try{
+            putDataList(FileRootTypes.index,default_sche_index_file_name,new ArrayList<>(scts));
+        }catch (Exception e){
+            check=false;
+            e.printStackTrace();
+        }
+        return check;
     }
 
     public void putRawSct(ScheWithTimeModel sct){
@@ -73,20 +80,15 @@ public class OftenOpts extends Fileystem {
     //获取已记录在案课表的所有课表名字
     public List<ScheWithTimeModel> getRecordedScts(){
         List<ScheWithTimeModel> scts=new ArrayList<>();
-
-        List<Object> lis=getDataList(FileRootTypes.index,default_sche_index_file_name,ScheWithTimeModel.class);
-        if(lis.size()>0)
-            CycleUtil.cycle(lis, (obj, objects) -> scts.add((ScheWithTimeModel) obj));
-
-
+        List<ScheWithTimeModel> lis=getDataList(FileRootTypes.index,default_sche_index_file_name,ScheWithTimeModel.class);
+        if(lis.size()>0&&lis.get(0)!=null) return lis;
         return scts;
     }
 
     //获取已记录在案课表的所有课表名字
     public List<Unimodel> getRecordedScheList(){
         List<Unimodel> unis_=new ArrayList<>();
-
-        List<Object> lis=getDataList(FileRootTypes.index,default_sche_index_file_name,ScheWithTimeModel.class);
+        List<ScheWithTimeModel> lis=getDataList(FileRootTypes.index,default_sche_index_file_name,ScheWithTimeModel.class);
         if(lis.size()>0&&lis.get(0)!=null){
             CycleUtil.cycle(lis, (obj, objects) -> {
                 ScheWithTimeModel sct=(ScheWithTimeModel)obj;
@@ -99,9 +101,9 @@ public class OftenOpts extends Fileystem {
     //快速获取索引信息,返回课表名字和时间课表名字，并对相关文件是否存在进行确认
     public ScheWithTimeModel getSctByIndex(int index){
         ScheWithTimeModel sct=new ScheWithTimeModel(0,default_sch,default_time);
-        List<Object> scts_=getDataList(FileRootTypes.index,default_sche_index_file_name,ScheWithTimeModel.class);
+        List<ScheWithTimeModel> scts_=getDataList(FileRootTypes.index,default_sche_index_file_name,ScheWithTimeModel.class);
         if(scts_.size()>0&&scts_.get(0)!=null){
-            ScheWithTimeModel sct_=(ScheWithTimeModel)scts_.get(index);
+            ScheWithTimeModel sct_=scts_.get(index);
             BasicOpts basicOpts=BasicOpts.getInstance(context);
             boolean isAccess=basicOpts.exist(FileRootTypes.sches,sct_.scheName)&&basicOpts.exist(FileRootTypes.times,sct_.timeName);
             if(isAccess) sct=sct_;
@@ -112,13 +114,11 @@ public class OftenOpts extends Fileystem {
     //通过索课表名字获取数据,返回课表名字和时间课表名字，并对相关文件是否存在进行确认
     public ScheWithTimeModel getSctByName(String scheName) {
         ScheWithTimeModel sct = new ScheWithTimeModel(0, default_sch, default_time);
-
-        List<Object> scts_ = getDataList(FileRootTypes.index, default_sche_index_file_name, ScheWithTimeModel.class);
+        List<ScheWithTimeModel> scts_ = getDataList(FileRootTypes.index, default_sche_index_file_name, ScheWithTimeModel.class);
 
         ScheWithTimeModel tc = null;
         if(scts_.size()>0&&scts_.get(0)!=null) {
-            for (Object obj : scts_) {
-                ScheWithTimeModel ct = (ScheWithTimeModel) obj;
+            for (ScheWithTimeModel ct : scts_) {
                 if (scheName.equals(ct.getScheName())) {
                     tc = ct;
                     break;
@@ -140,76 +140,25 @@ public class OftenOpts extends Fileystem {
     /*---作息表方法---*/
     public List<TimeModel> getRecordTms(){
         List<TimeModel> tms = new ArrayList<>();
-        List<Object> tms_=getDataList(FileRootTypes.times,default_time_index_file_name,TimeModel.class);
-        if(tms_.size()>0&&tms_.get(0)!=null)
-            CycleUtil.cycle(tms_, (obj, objects) -> tms.add((TimeModel) obj));
+        List<TimeModel> tms_=getDataList(FileRootTypes.times,default_time_index_file_name,TimeModel.class);
+        if(tms_.size()>0&&tms_.get(0)!=null) return tms_;
         return tms;
     }
 
+    /**
+     *
+     * @param time_name String
+     * @return Absolutely guarantee that ret will not be null object!!!
+     */
     public TimeHeadModel getThm(String time_name){
-        Object thm_=getData(FileRootTypes.times,time_name,TimeHeadModel.class);
-        if(thm_ instanceof String) return null;
-        else {
-            try{
-                TimeHeadModel thm=(TimeHeadModel) thm_;
-                thm=null;
-            }catch (Exception e){
-                thm_=null;
-                e.printStackTrace();
-            }
-            return thm_!=null?(TimeHeadModel)thm_:null;
-        }
+        TimeHeadModel thm_=getData(FileRootTypes.times,time_name,TimeHeadModel.class);
+        if(thm_ == null) return getThm(default_time+".txt");
+        return thm_;
     }
 
-    public Object getRawThm(String time_name){
-        return getData(FileRootTypes.times,time_name,TimeHeadModel.class);
+
+    public void putThms(List<TimeModel> thms){
+        if(thms==null||thms.size()==0) return;
+        putDataList(FileRootTypes.times,"time_index.txt",new ArrayList<>(thms));
     }
-
-    /* **
-     * 写入一个文件
-     * @param file_name 文件名
-     * @param thm 数据
-     *//*
-        public void putRawThm(String file_name,TimeHeadModel thm){
-            fileystem.putData(FileRootTypes.times,file_name,thm);
-        }*/
-
-    /*---作息表方法---*/
-/*
-    *//*-----课表作息表文件删除和重命名、索引表的添加和删除 START-----*//*
-    public boolean checkSctDocIsDuplicate(String oldName,String neoName){
-        List<ScheWithTimeModel> scts=getRecordedScts();
-        boolean isDuplicate=false;
-
-        if(oldName.equals(neoName)) return true;
-        if(scts.size()>0&&scts.get(0)!=null){
-            for (ScheWithTimeModel sct:scts) {
-                if(sct.getScheName().equals(neoName)){
-                    isDuplicate=true;
-                    break;
-                }
-            }
-        }
-
-        return isDuplicate;
-    }
-
-    public boolean checkThmDocIsDuplicate(String oldName,String neoName){
-        List<TimeModel> thms=getRecordTms();
-        boolean isDuplicate=false;
-        if(thms.size()>0&&thms.get(0)!=null){
-            if(oldName.equals(neoName)) return true;
-
-            for(TimeModel thm:thms){
-                if(thm.getTimeName().equals(neoName)){
-                    isDuplicate=true;
-                    break;
-                }
-            }
-        }
-
-        return isDuplicate;
-    }*/
-    /*-----课表作息表文件删除和重命名、索引表的添加和删除 END-----*/
-
 }
