@@ -4,7 +4,6 @@ import android.os.Bundle;
 
 import androidx.fragment.app.Fragment;
 
-import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -18,6 +17,7 @@ import com.hazukie.scheduleviews.activity.ScheCreateActivity;
 import com.hazukie.scheduleviews.fileutil.OftenOpts;
 import com.hazukie.scheduleviews.models.ClassLabel;
 import com.hazukie.scheduleviews.models.ScheWithTimeModel;
+import com.hazukie.scheduleviews.utils.CycleUtil;
 import com.hazukie.scheduleviews.utils.DisplayHelper;
 
 import java.util.List;
@@ -98,52 +98,37 @@ public class ScheFileCreateFrag extends Fragment {
         assert fragmentContainerAct != null;
         List<ClassLabel> main_list = fragmentContainerAct.getObjs();
         String globalTime = fragmentContainerAct.getParam();
-        Log.i("write>>>",""+ main_list.size()+",str="+ globalTime);
 
-        //FileHelper fileHelper=FileHelper.getInstance(getContext());
         try{
-            //获取所有课表名字，和其比较是否重复
-            List<ScheWithTimeModel> recorded_sches=oftenOpts.getRecordedScts();//FileHelper.getRecordedScts(getContext());
-            //List<ScheModel> recorded_sche=FileHelper.getRecordScms(getContext());
-            boolean isDuplicate=false;
-            if(recorded_sches.size()>0&&recorded_sches.get(0)!=null){
-                for(ScheWithTimeModel sche:recorded_sches){
-                    if (sche.scheName.equals(sche_file_name)) {
-                        isDuplicate = true;
-                        break;
-                    }
-                }
-            }
+            //获取所有课表名字
+            List<ScheWithTimeModel> recorded_sches=oftenOpts.getRecordedScts();
 
-           // Log.i("writInits: ","isDup="+(!isDuplicate)+", mains="+(main_list.size()>0));
-            if(!isDuplicate&& main_list.size()>0){
-                boolean isCreate= oftenOpts.putRawClsList(sche_file_name,main_list);//fileHelper.write(FileHelper.RootMode.sches,sche_file_name,new ArrayList<>(main_list));
+            if(main_list.size()>0){
+               boolean isCreate= oftenOpts.putRawClsList(sche_file_name,main_list);
+
+               //和其比较是否重复
                 recorded_sches.add(new ScheWithTimeModel(recorded_sches.size(),sche_file_name, globalTime));
+                CycleUtil.distinct(recorded_sches);
                 boolean isWrite2Index=oftenOpts.putRawSctzList(recorded_sches);
 
-                if(isCreate&&isWrite2Index){
-                    Toast.makeText(getContext(), "保存成功", Toast.LENGTH_SHORT).show();
-                    ScheCreateActivity.instance.finish();
-                    //关闭上一个活动！
-                    TimerTask task=new TimerTask() {
-                        @Override
-                        public void run() {
-                            getActivity().finish();
-                            Log.i("fragAct>>","closed!");
-                        }
-                    };
-                    Timer timer=new Timer();
-                    timer.schedule(task,300);
-
-                }else{
-                    Toast.makeText(getContext(), "保存失败！", Toast.LENGTH_SHORT).show();
-                }
-            }else if(main_list.size()==0){
-                DisplayHelper.Infost(getContext(),"您似乎没有进行添加操作！");
+               if(isCreate&&isWrite2Index){
+                   Toast.makeText(getContext(), "保存成功", Toast.LENGTH_SHORT).show();
+                   ScheCreateActivity.instance.finish();
+                   //关闭上一个活动！
+                   TimerTask task=new TimerTask() {
+                       @Override
+                       public void run() {
+                           requireActivity().finish();
+                       }
+                   };
+                   Timer timer=new Timer();
+                   timer.schedule(task,300);
+               }else{
+                   Toast.makeText(getContext(), "保存失败！", Toast.LENGTH_SHORT).show();
+               }
             }else {
-                DisplayHelper.Infost(getContext(),"文件已存在，请重新输入!");
+                DisplayHelper.Infost(getContext(), "您似乎没有进行添加操作！");
             }
-
         }catch (Exception e){
             e.printStackTrace();
         }
